Move header files in lib/Lower.
Make the directory structure more consistent and have all the header
files in one directory.
