+++ flang/lib/Lower/IntrinsicCall.cpp
@@ -14,12 +14,12 @@
 //===----------------------------------------------------------------------===//
 
 #include "flang/Lower/IntrinsicCall.h"
-#include "StatementContext.h"
-#include "SymbolMap.h"
 #include "flang/Common/static-multimap-view.h"
 #include "flang/Lower/ConvertType.h"
 #include "flang/Lower/Mangler.h"
 #include "flang/Lower/Runtime.h"
+#include "flang/Lower/StatementContext.h"
+#include "flang/Lower/SymbolMap.h"
 #include "flang/Lower/Todo.h"
 #include "flang/Optimizer/Builder/Character.h"
 #include "flang/Optimizer/Builder/Complex.h"
