+++ flang/lib/Lower/SymbolMap.cpp
@@ -10,8 +10,8 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "SymbolMap.h"
-#include "IterationSpace.h"
+#include "flang/Lower/SymbolMap.h"
+#include "flang/Lower/IterationSpace.h"
 #include "mlir/IR/BuiltinTypes.h"
 #include "llvm/Support/Debug.h"
